Abandoned using HTMLFormRenderer.
Screw that shit. No documentation/examples. Even after digging through the source code of the class itself it still does not render the fields. It produces the form tag with a submit button but it always say non_field_errors as the body. Will be moving to manual frontend building using javascript to control user input instead. We need to dynamically update stations based on systems and stuff so we need javascript and multiple request logic anyways.