Welp. Guess we going back to using django's templates. Also wrapper all api calls with a data field to solve the consistency vs object passing issue.
Why? Read some articles online about server vs client rendering. There were some good points about rendering the initial data on the server to decrease the wait time on the first render. I agrees with this approach. There are no point with having the client load a framework and then send another request to the server to ask for the data. Also the endpoints are mapped directly to our rest api anyways so the data is already available with no extra search and stuff.