We can now add connection to a route!!!!!!!!!!!!!
Various view and template changes to support javascript. jQuery is now used to override the input field with the url of the various model object when sending the post request. jQuery-ui is used to display the autocomplete information as well as using a callback to store the selected field to the form variable.