Completed #14, #11 and #5.
Tables are now used to show model properties with exception of list of complex models. Simple models such as StationCommodity was displayed in table as opposed to accordion. There is now a goto link on each panel if the panel is in a group. back_button also have been improved to provide more flexibility such as giving it an url, href or name to link against. It is possible to browse through all the data in the database without needing to manually modify the url.

The forms are now also collapsed by default. Since we already have the code for collapse there isn't really benefit in trying to change it to a modal. Some problem this might introduce is that there is now a lot of movement on the page because of the heavy use of collapse. We could retain the page structure by using modal to display all detailed information but that's a task and decision for another time.