Added detailed views for Commodity and Station to list the station_commodities list.
Still not sure on how to proceed to solve #16. We can either use client side js to paginate through the nested lists or we can link it to a search result(?) of the current model. Either way we would need to use some extension (Bootstrap table or drf-nested-routers) or some sort of a hack. Further investigation required.